+++ src/R_density.cpp
@@ -56,11 +56,9 @@ IntegerVector dbscan_density_int(
   }
 
   // cleanup
-  if(kdTree != NULL) {
-    delete kdTree;
-    annDeallocPts(dataPts);
-    annClose();
-  }
+  if (kdTree != NULL) delete kdTree;
+  if (dataPts != NULL)  annDeallocPts(dataPts);
+  // annClose(); is now done globaly in the package
 
   return count;
 }